impr: moved get_data_card endpoint to its own router
