+++ api/routers/annotation.py
@@ -10,7 +10,7 @@ sys.path.append(
     os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 )
 from src.text_search import text_var_search, vars_to_json, vars_dedup
-from src.connect import vars_formula_connection, dataset_header_document_dkg, vars_dataset_connection_simplified, construct_data_card
+from src.connect import vars_formula_connection, dataset_header_document_dkg, vars_dataset_connection_simplified
 from src.link_annos_to_pyacset import link_annos_to_pyacset
 
 router = APIRouter()
@@ -73,31 +73,6 @@ async def link_dataset_columns_to_dkg_info(gpt_key: str, csv_file: UploadFile = 
 
     if not success:
         return JSONResponse(status_code=status.HTTP_401_UNAUTHORIZED, content=s)
-
-    return ast.literal_eval(s)
-
-
-@router.post("/get_data_card", tags=["Paper-2-annotated-vars"])
-async def get_data_card(gpt_key: str, csv_file: UploadFile = File(...), doc_file: UploadFile = File(...)):
-
-    csv_string = await csv_file.read()
-    csv_string = csv_string.decode()
-    csv_strings = csv_string.split('\n')
-
-    if len(csv_strings) == 0:
-        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST, content="Empty CSV file")
-
-    # get header and <=5 random rows
-    num_rows_to_sample = min(5, len(csv_strings) - 1)
-    csv_str = csv_strings[0] + '\n' + '\n'.join(random.sample(csv_strings[1:], num_rows_to_sample))
-
-    doc = await doc_file.read()
-    doc = doc.decode()
-
-    s, success = construct_data_card(data=csv_str, data_doc=doc, gpt_key=gpt_key)
-
-    if not success:
-        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST, content=s)
 
     return ast.literal_eval(s)
